+++ CHANGELOG.md
@@ -5,4 +5,5 @@ Change logs are organized by package:
 - [@ariakit/core](packages/ariakit-core/CHANGELOG.md)
 - [@ariakit/react](packages/ariakit-react/CHANGELOG.md)
 - [@ariakit/react-core](packages/ariakit-react-core/CHANGELOG.md)
+- [@ariakit/solid-core](packages/ariakit-solid-core/CHANGELOG.md)
 - [@ariakit/test](packages/ariakit-test/CHANGELOG.md)